Self Made Basic Notification
Fake it till you make it!

Created one function in notificationSelfmade.js that shows the html notification (in index2). You'll see it by clicking on "Show notification". If you click at the picture the Framework7 notification still appears. The idea is to insert the different content items depending on what notification is shown. That's why the notification BuyCall is shown as an object. Still experimenting with it though. Swiping isn't implemented yet.